A first draft of a Vob.
The interface is designed to be a superset of Vec<bool>'s interface. Several
functions are missing, because it would be way too much to do it all in one go.
However, this contains all the functionality I need, and several things I don't,
so I think it's a reasonable start.

There's extensive documentation in the commit, intentionally based on Vec's
documentation whenever possible.