Make iter_set_bits & iter_unset_bits take range arguments.
We will never be able to take a slice of a Vob, because a slice needs to be a
(byte aligned) pointer. That means that the previous definitions of
iter_set_bits and iter_unset_bits were only useful if you wanted to find either
a) every set/unset bit b) the first N set/unset bits.

In particular, you could not say things like "find me all set/unset bits between
positions 50 and 100". iter_set_bits().skip(50).take(50) might seem like it
would work, but this would take the first 50 set bits, none of which might be
between positions 50 and 100! Instead, one had to manually check each bit which
is very slow.

This commit changes that: iter_set_bits and iter_unset_bits now take a range,
like "drain". You can say things like "iter_set_bits(..)" or
"iter_set_bits(10..)" and so on.

The bad news is that RangeArgument is unstable; that means we have to have our
own version of the API in this crate. Since it's expected that Rust will
eventually call this RangeBounds, I've chosen to use that terminology. This is a
temporaryish hack until this feature is eventually stabilised in Rust, but it
means that we can still run things on stable in the meantime.