crates.io still doesn't know about UPL-1.0.
Even though the actual license is Apache/MIT/UPL, we pretend to crates.io that
it's just Apache/MIT. When, eventually, crates.io picks up the UPL, we can add
that back in as another "OR". But I think simply dropping it from Cargo.toml is
better than having the license show up as "non-standard".