Fix example in README.md and also slightly expand it.
The previous example had an incorrect "assert" (that compared an Option against
an int). Fix that, and tweak the example so that we can sensibly show
iter_set_bits. It turns out that this function is one of Vob's key performance
advantages over other bit vector libraries, so it makes sense to include that in
the README IMHO.