Make set() panic if the index is out of bounds.
The original idea was to make set() have the same interface as get(), which
returns None if the index is out of bounds. However, this wasn't a great choice,
because users always check the return value of get() whereas most people ignore
the return value of set(). Thus out of bounds set() calls tended to silently be
ignored.