Move to Criterion for benchmarking.
`cargo bench` is still a nightly feature, which makes it easy not to
notice that changes have broken benchmarking (my desktop, for example,
exclusively runs stable).

This commit moves us from Rust's in-built benchmarking system to
Criterion. This increases the boiler-plate but a) it works on stable (so
we can now also test in CI!) b) it provides better statistical output.